📝 docs(README.md): add sections for Transaction and ODM Integration 🔥 remove(user.dto.ts): remove email field from UserDto ✨ feat(user.module.ts): add UserProfile to providers 🆕 add(user.profile.ts): create new UserProfile class for automapper 🔧 refactor(user.schema.ts): add AutoMap decorators and set mongoose 'strictQuery' option
The README.md file was updated to include sections for Transaction and ODM Integration, providing more information about these aspects of the project. The email field was removed from UserDto as it was deemed unnecessary. UserProfile was added to the providers in user.module.ts to enable the use of automapper. A new UserProfile class was created in user.profile.ts for the same purpose. In user.schema.ts, AutoMap decorators were added to the schema properties to enable automatic mapping, and the 'strictQuery' option was set in mongoose for stricter query validation. This was done to improve data integrity and validation in the application.