🔧 refactor(master_tired-potato.yml): remove Azure docs comments, add zip command, change artifact path and package to node-app.zip 🔥 remove(BaseService.ts): remove unused import 'class-transformer' from BaseService.ts
The Azure docs comments were removed from the workflow file as they were not necessary for the workflow's functionality. The 'zip' command was added to compress the Node.js app into a zip file before uploading it as an artifact. This change was made to reduce the size of the artifact, making the upload process faster and more efficient. The path of the artifact and the package in the deployment job were changed to 'node-app.zip' to reflect this change.

The 'class-transformer' import was removed from BaseService.ts as it was not being used, which helps to keep the codebase clean and efficient.