✨ feat(app.module.ts): import FormSubmissionsModule and FormBuilderModule 🔧 refactor(form-builder.entity.ts, form-submission.entity.ts): import SchemaFactory and create schema for FormBuilder and FormSubmission 🔥 remove(form-builder.controller.spec.ts, form-builder.service.spec.ts, form-submissions.controller.spec.ts, form-submissions.service.spec.ts): delete unnecessary test files 🔧 refactor(form-builder.module.ts, form-submissions.module.ts): import MongooseModule and add FormBuilder and FormSubmission to MongooseModule.forFeature
The FormSubmissionsModule and FormBuilderModule were imported into app.module.ts to extend the functionality of the application. The SchemaFactory was imported and used to create schemas for FormBuilder and FormSubmission entities, which will allow for easier
manipulation of these entities in the database. Unnecessary test files were removed to clean up the codebase. The MongooseModule was imported and FormBuilder and FormSubmission were added to MongooseModule.forFeature in their respective modules to enable the use of these entities with Mongoose.