🔧 refactor(master_tired-potato.yml): rename zip file from node-app.zip to nest.zip for better clarity ➕ feat(master_tired-potato.yml): add step to delete zip file after deployment to clean up workspace
The zip file was renamed from node-app.zip to nest.zip to better reflect the contents of the file, which is a NestJS application. This change improves clarity and understanding of the file's purpose. Additionally, a new step was added to delete the zip file after deployment. This is to ensure that the workspace remains clean and uncluttered after the deployment process.